Move FilterHopperMenu to new menu folder
